atelet: move the sparse file copy helpers into internal/sparsefile
copyFile and its hole-preserving machinery lived in package main, usable
only by atelet's own checkpoint staging. The filecache package is about
to need the same copy (its copy-out mode hands consumers a private,
hole-preserved copy of a cached artifact), so move the code where both
can import it.

Mechanical move, with one seam added: Copy(src, dst *os.File) exposes
the engine on caller-owned handles, for callers that must open the
source before its name can vanish or create the destination with
O_EXCL. CopyFile keeps its os.Create semantics for the existing caller.
