+++ cmd/atelet/main_test.go
@@ -341,63 +341,6 @@ func TestWriteFileAtomic(t *testing.T) {
 	})
 }
 
-func TestCopyFile(t *testing.T) {
-	dir := t.TempDir()
-	src := filepath.Join(dir, "src")
-	want := []byte("checkpoint pages")
-	if err := os.WriteFile(src, want, 0o600); err != nil {
-		t.Fatalf("seeding src: %v", err)
-	}
-
-	dst := filepath.Join(dir, "dst")
-	n, err := copyFile(src, dst)
-	if err != nil {
-		t.Fatalf("copyFile: %v", err)
-	}
-	if n != int64(len(want)) {
-		t.Errorf("copied %d bytes, want %d", n, len(want))
-	}
-	got, err := os.ReadFile(dst)
-	if err != nil {
-		t.Fatalf("reading dst: %v", err)
-	}
-	if !bytes.Equal(got, want) {
-		t.Errorf("dst content = %q, want %q", got, want)
-	}
-
-	if _, err := copyFile(dir, filepath.Join(dir, "dst2")); err == nil {
-		t.Error("copyFile(directory, ...) succeeded, want error")
-	}
-}
-
-type failingCloseFile struct{ *os.File }
-
-func (f failingCloseFile) Close() error {
-	_ = f.File.Close()
-	return errors.New("deferred flush failed")
-}
-
-func TestCopyFile_CloseError(t *testing.T) {
-	orig := createDestFile
-	createDestFile = func(name string) (io.WriteCloser, error) {
-		f, err := os.Create(name)
-		if err != nil {
-			return nil, err
-		}
-		return failingCloseFile{f}, nil
-	}
-	t.Cleanup(func() { createDestFile = orig })
-
-	dir := t.TempDir()
-	src := filepath.Join(dir, "src")
-	if err := os.WriteFile(src, []byte("checkpoint pages"), 0o600); err != nil {
-		t.Fatalf("seeding src: %v", err)
-	}
-	if _, err := copyFile(src, filepath.Join(dir, "dst")); err == nil {
-		t.Error("copyFile with failing destination Close = nil, want error")
-	}
-}
-
 // validRunRequest, validCheckpointRequest, and validRestoreRequest build
 // requests whose every field passes validation; the per-request tests below
 // break one field per case.
@@ -1393,175 +1336,6 @@ func TestBuildAteomWorkloadSpec_ImageVolumeMounts(t *testing.T) {
 	}
 	if len(ctr.GetDurableDirVolumeMounts()) != 1 || ctr.GetDurableDirVolumeMounts()[0].GetVolumeName() != "data" {
 		t.Errorf("durable mounts = %v, want just data", ctr.GetDurableDirVolumeMounts())
-	}
-}
-
-// allocatedBytes reports how much disk a file actually occupies, which is less than its
-// size when it has holes.
-func allocatedBytes(t *testing.T, path string) int64 {
-	t.Helper()
-	var st syscall.Stat_t
-	if err := syscall.Stat(path, &st); err != nil {
-		t.Fatalf("stat %q: %v", path, err)
-	}
-	return st.Blocks * 512
-}
-
-// noFdFile hides the descriptor of an *os.File, forcing copySparse down its
-// userspace path.
-type noFdFile struct {
-	f *os.File
-}
-
-func (n noFdFile) Write(b []byte) (int, error)              { return n.f.Write(b) }
-func (n noFdFile) WriteAt(b []byte, off int64) (int, error) { return n.f.WriteAt(b, off) }
-func (n noFdFile) Truncate(size int64) error                { return n.f.Truncate(size) }
-func (n noFdFile) Close() error                             { return n.f.Close() }
-
-func TestCopyFilePreservesHoles(t *testing.T) {
-	const (
-		size     = 32 << 20
-		markerAt = 16 << 20
-	)
-	dir := t.TempDir()
-	src := filepath.Join(dir, "memory-ranges")
-
-	// A stand-in for a guest memory image: mostly hole, with data at both the start
-	// and the middle.
-	f, err := os.Create(src)
-	if err != nil {
-		t.Fatalf("creating src: %v", err)
-	}
-	if err := f.Truncate(size); err != nil {
-		t.Fatalf("sizing src: %v", err)
-	}
-	head := bytes.Repeat([]byte{0xAB}, 4<<10)
-	middle := bytes.Repeat([]byte{0xCD}, 4<<10)
-	if _, err := f.WriteAt(head, 0); err != nil {
-		t.Fatalf("writing head: %v", err)
-	}
-	if _, err := f.WriteAt(middle, markerAt); err != nil {
-		t.Fatalf("writing middle: %v", err)
-	}
-	if err := errors.Join(f.Sync(), f.Close()); err != nil {
-		t.Fatalf("flushing src: %v", err)
-	}
-
-	dst := filepath.Join(dir, "copied")
-	n, err := copyFile(src, dst)
-	if err != nil {
-		t.Fatalf("copyFile: %v", err)
-	}
-	if n != size {
-		t.Errorf("copied %d logical bytes, want %d", n, size)
-	}
-
-	// The copy must be byte-identical, holes included.
-	want, err := os.ReadFile(src)
-	if err != nil {
-		t.Fatalf("reading src: %v", err)
-	}
-	got, err := os.ReadFile(dst)
-	if err != nil {
-		t.Fatalf("reading dst: %v", err)
-	}
-	if !bytes.Equal(want, got) {
-		t.Fatal("copy differs from source")
-	}
-
-	srcAlloc, dstAlloc := allocatedBytes(t, src), allocatedBytes(t, dst)
-	if srcAlloc >= size/2 {
-		t.Skipf("source did not end up sparse (%d of %d bytes allocated); "+
-			"this filesystem cannot report holes", srcAlloc, int64(size))
-	}
-	// A dense copy would allocate the full logical size; a hole-preserving one stays
-	// near the source's footprint.
-	if dstAlloc > srcAlloc*4 {
-		t.Errorf("copy allocated %d bytes for a %d-byte source (logical %d): holes were filled in",
-			dstAlloc, srcAlloc, int64(size))
-	}
-}
-
-func TestCopyFileAllHoles(t *testing.T) {
-	const size = 8 << 20
-	dir := t.TempDir()
-	src := filepath.Join(dir, "empty")
-	f, err := os.Create(src)
-	if err != nil {
-		t.Fatalf("creating src: %v", err)
-	}
-	if err := errors.Join(f.Truncate(size), f.Close()); err != nil {
-		t.Fatalf("sizing src: %v", err)
-	}
-
-	dst := filepath.Join(dir, "copied")
-	if _, err := copyFile(src, dst); err != nil {
-		t.Fatalf("copyFile: %v", err)
-	}
-	st, err := os.Stat(dst)
-	if err != nil {
-		t.Fatalf("stat dst: %v", err)
-	}
-	if st.Size() != size {
-		t.Errorf("copy is %d bytes, want %d", st.Size(), int64(size))
-	}
-}
-
-// TestCopyFilePreservesHolesUserspace covers the fallback taken when the destination
-// does not expose a descriptor, so copy_file_range is unavailable.
-func TestCopyFilePreservesHolesUserspace(t *testing.T) {
-	orig := createDestFile
-	createDestFile = func(name string) (io.WriteCloser, error) {
-		f, err := os.Create(name)
-		if err != nil {
-			return nil, err
-		}
-		return noFdFile{f: f}, nil
-	}
-	t.Cleanup(func() { createDestFile = orig })
-
-	const size = 32 << 20
-	dir := t.TempDir()
-	src := filepath.Join(dir, "memory-ranges")
-	f, err := os.Create(src)
-	if err != nil {
-		t.Fatalf("creating src: %v", err)
-	}
-	if err := f.Truncate(size); err != nil {
-		t.Fatalf("sizing src: %v", err)
-	}
-	marker := bytes.Repeat([]byte{0xEF}, 4<<10)
-	if _, err := f.WriteAt(marker, 8<<20); err != nil {
-		t.Fatalf("writing marker: %v", err)
-	}
-	if err := errors.Join(f.Sync(), f.Close()); err != nil {
-		t.Fatalf("flushing src: %v", err)
-	}
-
-	dst := filepath.Join(dir, "copied")
-	if _, err := copyFile(src, dst); err != nil {
-		t.Fatalf("copyFile: %v", err)
-	}
-
-	want, err := os.ReadFile(src)
-	if err != nil {
-		t.Fatalf("reading src: %v", err)
-	}
-	got, err := os.ReadFile(dst)
-	if err != nil {
-		t.Fatalf("reading dst: %v", err)
-	}
-	if !bytes.Equal(want, got) {
-		t.Fatal("userspace copy differs from source")
-	}
-
-	srcAlloc, dstAlloc := allocatedBytes(t, src), allocatedBytes(t, dst)
-	if srcAlloc >= size/2 {
-		t.Skipf("source did not end up sparse (%d of %d bytes allocated)", srcAlloc, int64(size))
-	}
-	if dstAlloc > srcAlloc*4 {
-		t.Errorf("userspace copy allocated %d bytes for a %d-byte source: holes were filled in",
-			dstAlloc, srcAlloc)
 	}
 }
